[Feat] 지도 테그 선택뷰 로직 변경
모두보기를 누를시 다른 다른 테그들의 active상태가 초기화 된다.
모두보기를 누른 상태에서 모두보기를 다시눌러도 모두보기 상태가 취소되지 않는다.
해당 상태에서 다른 테그를 누를 수 모두보기가 해제된다.
선택한 테그만 active상태가 된다.
위 기능들은 SelectMapTagViewModel타입을 생성하여 관리한다.
